Rename viz, remove references to X
