+++ Makefile
@@ -137,13 +137,5 @@ test installcheck: install_test
 
 test_all: install_test
 	$(MAKE) -C $(BUILD_DIR)/test --no-print-directory run_ctest_all
-
-test_all_no_x: install_test    
-	cd build/installed/bin/; \
-	./gtester  --gtest_filter=-*_x  --gtest_output=xml:gtester_all_no_x.xml
-
-test_only_x: install_test    
-	cd build/installed/bin/; \
-	./gtester  --gtest_filter=*_x  --gtest_output=xml:gtester_only_x.xml
 	
 #############################################################################
