+++ src/main/cpp/http/server.h
@@ -1,11 +1,12 @@
-#ifndef VISUALISER_H_INCLUDED
-#define VISUALISER_H_INCLUDED
+#ifndef STRIDEHTTPSERVER_H_INCLUDED
+#define STRIDEHTTPSERVER_H_INCLUDED
 
 #include <unistd.h>
 #include <iostream>
 
 
 namespace Stride {
+namespace HTTP {
 
 class Visualiser{
     public:
@@ -13,6 +14,7 @@ class Visualiser{
     void run();
 };
 
+}
 }
 
 #endif  // end of include guard
