+++ src/test/cpp/gtester/CMakeLists.txt
@@ -31,7 +31,7 @@ set( SRC
 )
 
 if( Poco_FOUND )
-	set( SRC ${SRC} VizProtoTest.cpp)
+	set( SRC ${SRC} HTTPProtoTest.cpp)
 endif()
 
 add_executable(${EXEC}   ${SRC} $<TARGET_OBJECTS:libstride> $<TARGET_OBJECTS:trng>)
@@ -51,12 +51,6 @@ add_test( NAME  ${EXEC}_all
         WORKING_DIRECTORY  ${TESTS_DIR}
 		COMMAND   ${CMAKE_INSTALL_PREFIX}/${BIN_INSTALL_LOCATION}/${EXEC}
 		    --gtest_output=xml:gtester_all.xml
-)
-
-add_test( NAME  ${EXEC}_all_no_x
-        WORKING_DIRECTORY  ${TESTS_DIR}
-		COMMAND   ${CMAKE_INSTALL_PREFIX}/${BIN_INSTALL_LOCATION}/${EXEC}
-		    --gtest_filter=-*_x  --gtest_output=xml:gtester_all_no_x.xml
 )
 
 #============================================================================
